multitask_recommender/requirements.txt
tensorflow
tensorflow-recommenders
tensorflow-datasets
pyyaml
